requirements.txt
pydub==0.25.1
python_telegram_bot==20.0b0
python-telegram-bot[all]==20.0b0
whisper @ git+https://github.com/openai/whisper.git
replicate==0.4.0